+++ src/cpp/impl/CellSim.Cells.CellBehaviorPtr.hpp
@@ -31,7 +31,7 @@ namespace CellSim::Cells
         [[nodiscard]] static CellBehaviorPtr FromPointer(CellBehavior* ptr);
 
         /// @brief nullptr
-        [[nodiscard]] static constexpr CellBehaviorPtr Null() noexcept;
+        [[nodiscard]] static CellBehaviorPtr Null() noexcept;
 
         CellBehaviorPtr(CellBehaviorPtr const& right);
         constexpr CellBehaviorPtr(CellBehaviorPtr&& right) noexcept;
@@ -69,7 +69,7 @@ namespace CellSim::Cells
         delete m_ptr;
     }
 
-    constexpr CellBehaviorPtr CellBehaviorPtr::Null() noexcept
+    inline CellBehaviorPtr CellBehaviorPtr::Null() noexcept
     {
         return CellBehaviorPtr((CellBehavior*)nullptr);
     }