+++ src/cpp/impl/CellSim.Graphics.Color.hpp
@@ -2,6 +2,7 @@
 #define CELLSIM_GRAPHICS_COLOR_HPP
 
 #include "base.hpp"
+#include <string_view>
 
 namespace CellSim::Graphics
 {
@@ -12,6 +13,8 @@ namespace CellSim::Graphics
         uint8_t G;
         uint8_t B;
 
+        Color() = default;
+        explicit Color(::std::string_view value);
         constexpr Color(uint8_t r, uint8_t g, uint8_t b) noexcept;
         constexpr Color(uint8_t a, uint8_t r, uint8_t g, uint8_t b) noexcept;
     };
@@ -19,6 +22,8 @@ namespace CellSim::Graphics
 
 namespace CellSim::Graphics
 {
+    
+
     constexpr Color::Color(uint8_t r, uint8_t g, uint8_t b) noexcept
         : A(255)
         , R(r)