+++ src/cpp/impl/CellSim.Settings.Config.CellBehavior.cpp
@@ -13,21 +13,6 @@ namespace CellSim::Settings
 {
     void Config::CellBehavior::Load(::nlohmann::json const& config)
     {
-        ::std::string s;
-
-        try {
-            s = config.at("behaviorType").get<::std::string>();
-        }
-        catch (...) {
-            throw ::std::runtime_error(Messages::Get("Settings.Config.CellBehavior.Load.Error.JsonError"));
-        }
-
-        if (s == "MoleculeAware") s_behaviorType = Cells::CellBehaviorType::MoleculeAware;
-        else if (s == "Normal") s_behaviorType = Cells::CellBehaviorType::Normal;
-        else if (s == "User") s_behaviorType = Cells::CellBehaviorType::User;
-        else if (s == "WavePropagation") s_behaviorType = Cells::CellBehaviorType::WavePropagation;
-        else [[unlikely]] throw ::std::runtime_error(Messages::Get("Settings.Config.CellBehavior.Load.Error.behaviorType"));
-
         MoleculeAware::Load(config["moleculeAware"]);
         Normal::Load(config["normal"]);
         User::Load(config["user"]);